requirements.txt
discord
psutil
discord-pretty-help
jishaku